+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Not released
 
+- Support for `onRowMouseEnter` and `onRowMouseLeave` handlers for Table Widget [#907](https://github.com/CartoDB/carto-react/pull/907)
 
 ## 3.0.0
 
 ### 3.0.0-alpha.20 (2024-09-20)
+
 - `NoWrap` for table header [#908](https://github.com/CartoDB/carto-react/pull/908)
 
 ### 3.0.0-alpha.19 (2024-09-19)